// src/shared/api/generated/model/importClientsRequestDtoRowsItem.ts
/**
 * Generated by orval 🍺
 * Do not edit manually.
 * Yoclick API
 * API multi-centro de reservas en marca blanca. Contrato canónico para la app.
 * OpenAPI spec version: 0.1.0
 */
import type { ImportClientsRequestDtoRowsItemLevel } from './importClientsRequestDtoRowsItemLevel.ts';

export type ImportClientsRequestDtoRowsItem = {
  /**
     * @minLength 1
     * @maxLength 120
     */
  fullName: string;
  /**
     * @maxLength 254
     * @nullable
     * @pattern ^(?:[A-Za-z0-9_'+\-]+\.)*[A-Za-z0-9_'+\-]*[A-Za-z0-9_+-]@(?:[A-Za-z0-9][A-Za-z0-9\-]*\.)+[A-Za-z]{2,}$
     */
  email?: string | null;
  /**
     * @maxLength 30
     * @nullable
     */
  phone?: string | null;
  /** @nullable */
  level?: ImportClientsRequestDtoRowsItemLevel;
};