// src/shared/api/generated/model/importClientsRequestDtoRowsItemLevel.ts
/**
 * Generated by orval 🍺
 * Do not edit manually.
 * Yoclick API
 * API multi-centro de reservas en marca blanca. Contrato canónico para la app.
 * OpenAPI spec version: 0.1.0
 */

/**
 * @nullable
 */
export type ImportClientsRequestDtoRowsItemLevel = typeof ImportClientsRequestDtoRowsItemLevel[keyof typeof ImportClientsRequestDtoRowsItemLevel] | null;


export const ImportClientsRequestDtoRowsItemLevel = {
  beginner: 'beginner',
  intermediate: 'intermediate',
  advanced: 'advanced',
} as const;